Hack to avoid negative durations
Rarely, the next interval computed is too far in the future. When this happens,
now < next, and the elapsed time calcdiff(now, next) is negative. If such a
situation occurs, it is obvious the calculation of the next interval was wrong,
and we rewind it until we find a positive value.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>