Fix cyclictest negative durations
When the next interval is too far in the future, then now < next, and the
elapsed time calcdiff(now, next) is negative. If such a situation occurs, it is
obvious the calculation of the next interval was wrong, and we rewind it until
we find a positive value.

This bug can be reproduced by setting a very short interval, such as 1us:

Without the patch:
  ./cyclictest -i 1
  T: 0 ( 8429) P: 0 I:1 C:1099294 Min:      0 Act:  -34 Avg:-9223372036854775808 Max:      -1

With the patch:
  ./cyclictest -i 1
  T: 0 ( 8665) P: 0 I:1 C: 419151 Min:      0 Act:    0 Avg:    0 Max:      76

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>